Through testing, found that limiting TSP solving time to 1 second yields almost as good results
It finds a lot of results in a short period of time, and then tends to take exponentially longer time for very minimal gains. Limiting execution time to 1 second (actually much less) compared to a 17 second solve resulted in a loss of 7 ly total, which is a rounding error more or less.